Fix reset of checlist when switching view modes
If a user reset a checklist in reading mode and then quickly switched to
edit mode, the state of the checlist would reset. Essentialy what was
happening was that in read mode the file was not syncing quickly enough.
To solve this we save the document directly after reseting the
checklist.